cleaning up disk angular momentum multipplot

+++ FLASH_analysis/inner_disk_multi.py
@@ -188,7 +188,7 @@ for spin_lab in Spin_labels:
             Time_array, Radius_array, All_profiles_array, Total_inner_disk = pickle.load(file)
             file.close()
             
-            Radius_array = np.array(Radius_array)
+            Radius_array = np.array(Radius_array).astype(float)
             Radius_array[np.where(Radius_array==100)] = np.nan
             
             ax2 = axs.flatten()[plot_it].twinx()
